Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
sentence-transformers
faiss-cpu
groq
python-multipart
pillow
python-dotenv
pytesseract
PyPDF2
alembic
httpx